+++ app/src/main/java/org/mozilla/tv/firefox/telemetry/TelemetryIntegration.kt
@@ -112,7 +112,6 @@ open class TelemetryIntegration protected constructor(
         const val FXA_SHOW_PROFILE_BUTTON = "fxa_show_profile_button"
         const val FXA_GET_TABS_BUTTON = "fxa_get_tabs_button"
         const val FXA_SIGN_OUT_BUTTON = "fxa_sign_out_button"
-        const val FXA_GO_BACK_BUTTON = "fxa_go_back_button"
     }
 
     private object Extra {
@@ -301,16 +300,12 @@ open class TelemetryIntegration protected constructor(
         TelemetryEvent.create(Category.ACTION, Method.CLICK, Object.FXA, Value.FXA_SHOW_PROFILE_BUTTON).queue()
     }
 
-    fun fxaProfileGetTabsButtonClickEvent() {
+    fun fxaProfileShowOnboardingButtonClickEvent() {
         TelemetryEvent.create(Category.ACTION, Method.CLICK, Object.FXA, Value.FXA_GET_TABS_BUTTON).queue()
     }
 
     fun fxaProfileSignOutButtonClickEvent() {
         TelemetryEvent.create(Category.ACTION, Method.CLICK, Object.FXA, Value.FXA_SIGN_OUT_BUTTON).queue()
-    }
-
-    fun fxaProfileGoBackButtonClickEvent() {
-        TelemetryEvent.create(Category.ACTION, Method.CLICK, Object.FXA, Value.FXA_GO_BACK_BUTTON).queue()
     }
 
     /**