+++ app/src/test/java/org/mozilla/tv/firefox/fxa/FxaRepoTest.kt
@@ -143,7 +143,7 @@ class FxaRepoTest {
     fun `GIVEN profile has no avatar WHEN profile is converted to domain object THEN default resource image should be used`() {
         val profile = Profile("uid", null, null, null)
 
-        assertEquals(defaultProfileAvatarImage, profile.toDomainObject().avatar)
+        assertEquals(defaultProfileAvatarImage, profile.toDomainObject().avatarSetStrategy)
     }
 
     @Test
@@ -156,7 +156,7 @@ class FxaRepoTest {
         )
 
         profiles.map { it.toDomainObject() }
-            .map { it.avatar }
+            .map { it.avatarSetStrategy }
             .forEach { assertEquals(defaultProfileAvatarImage, it) }
     }
 
@@ -172,7 +172,7 @@ class FxaRepoTest {
         )
 
         profiles.map { it.toDomainObject() }
-            .map { it.avatar }
+            .map { it.avatarSetStrategy }
             .forEach { assertEquals(expectedStrategy, it) }
     }
 }